Built first POST interaction and full rounded API
Able to POST data then GET it from the view, first step in creating the
dynamic form. So far the template for the form session data is created
for each student (not family, etc...) - I also added the favicon.ico to
the local folder. Added some functions to the API for sanitizing inputs
both names of form inputs and types. Good commit! :+1: